requirements.txt
streamlit
seaborn
matplotlib
pandas
mysql-connector-python